+++ test/integration/production_release/mishka_prod_test/lib/mishka_prod_test/durable_plugin.ex
@@ -1,0 +1,7 @@
+defmodule MishkaProdTest.DurablePlugin do
+  @moduledoc false
+  use MishkaInstaller.Event.Hook, event: "prod_durable_evt", initial: %{priority: 10}
+
+  @impl true
+  def call(state), do: {:reply, state}
+end